feat(mobile): resend the login form page
For an unknow reason, my last version of the login wasn't push on the repository. So I'll do it RN. I really don't know why lmao
