+++ mobile/lib/app/features/login/view/login_screen.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:triggo/app/features/login/bloc/login_bloc.dart';
-import 'package:triggo/app/features/login/view/email.screen.dart';
-import 'package:triggo/app/features/login/view/password.screen.dart';
+import 'package:triggo/app/features/login/view/login_form.dart';
 import 'package:triggo/mediator/authentication.mediator.dart';
 
 class LoginScreen extends StatelessWidget {
@@ -14,34 +13,23 @@ class LoginScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final authenticationMediator =
-        RepositoryProvider.of<AuthenticationMediator>(context);
-
-    return BlocProvider(
-      create: (context) =>
-          LoginBloc(authenticationMediator: authenticationMediator),
-      child: const _LoginNavigator(),
-    );
-  }
-}
-
-class _LoginNavigator extends StatelessWidget {
-  const _LoginNavigator();
-
-  @override
-  Widget build(BuildContext context) {
-    return Navigator(
-      onGenerateRoute: (settings) {
-        switch (settings.name) {
-          case '/email':
-            return MaterialPageRoute(builder: (_) => const EmailInputScreen());
-          case '/password':
-            return MaterialPageRoute(
-                builder: (_) => const PasswordInputScreen());
-          default:
-            return MaterialPageRoute(builder: (_) => const EmailInputScreen());
-        }
-      },
+    return Scaffold(
+      appBar: AppBar(
+        leading: IconButton(
+          icon: const Icon(Icons.arrow_back),
+          onPressed: () => Navigator.of(context).pop(),
+        ),
+      ),
+      extendBodyBehindAppBar: true,
+      body: Padding(
+        padding: const EdgeInsets.all(12),
+        child: BlocProvider(
+          create: (context) => LoginBloc(
+            authenticationMediator: context.read<AuthenticationMediator>(),
+          ),
+          child: const LoginForm(),
+        ),
+      ),
     );
   }
 }
